+++ cmd/oci-cloud-controller-manager/main.go
@@ -21,17 +21,15 @@ import (
 	"os"
 	"time"
 
+	_ "github.com/oracle/oci-cloud-controller-manager/pkg/cloudprovider/providers/oci"
+	"github.com/oracle/oci-cloud-controller-manager/pkg/logging"
 	"github.com/spf13/pflag"
 	"go.uber.org/zap"
-
 	utilflag "k8s.io/apiserver/pkg/util/flag"
 	"k8s.io/apiserver/pkg/util/logs"
 	"k8s.io/kubernetes/cmd/cloud-controller-manager/app"
 	_ "k8s.io/kubernetes/pkg/client/metrics/prometheus" // for client metric registration
 	_ "k8s.io/kubernetes/pkg/version/prometheus"        // for version metric registration
-
-	"github.com/oracle/oci-cloud-controller-manager/pkg/logging"
-	_ "github.com/oracle/oci-cloud-controller-manager/pkg/oci"
 )
 
 var version string
