+++ pkg/cloudprovider/providers/oci/config_validate.go
