+++ pkg/cloudprovider/providers/oci/config_validate_test.go
