+++ pkg/oci/instance/metadata/instance_metadata_mock.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package instancemeta
+package metadata
 
 type mockMetadataGetter struct {
 	metadata *InstanceMetadata
